+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -186,7 +186,7 @@ class Data(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-27T08:59:46.066Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-27T09:34:54.012Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")