+++ packages/orq-rc/RELEASES.md
@@ -3639,3 +3639,13 @@ Based on:
 - [python v3.14.0-rc.59] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.59] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.59 - packages/orq-rc
+
+## 2025-10-27 09:36:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.3 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.60] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.60] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.60 - packages/orq-rc